Add better support for static variables (add enums, more initializer types); add integer modifiers (i.e. 24L)

+++ src/compiler/ast.h
@@ -106,7 +106,7 @@ typedef struct {
 
     EagleBasicType etype;
     union {
-        int32_t i;
+        long long i;
         double d;
         char *id;
     } value;
